requirements.txt
aenum==3.1.15
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contentful-management==2.11.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-appconf==1.0.5
django-compressor==4.4
django-cors-headers==4.2.0
django-filter==2.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.21.6
ecdsa==0.18.0
flatdict==4.0.1
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
kombu==5.3.1
multidict==6.0.4
oauthlib==3.2.2
okta==2.9.2
packaging==23.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pycryptodomex==3.18.0
pydash==7.0.5
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
rcssmin==1.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
xmltodict==0.13.0
yarl==1.9.2
pgvector
sentence_transformers
whitenoise==6.5.0
openai
langchain
pypdf
docx2txt
tiktoken